Oops: undead disposables + `DisposeState` redesign
A disposed object should never revert back to being undisposed, otherwise it
would appear that they're operable again when that's very likely not the case.

This fix adds a new state that indicates that an object is disposed even though
not fully disposed, and we revert to this new state instead of reverting to a
state of not being disposed.