Simplified instead
- Almost all `Get…()` methods, with a few exceptions, should simply
return either the decoded value or a default value, without casting the
data bytes by force if the type hint is different.
- For the `Text` type, we'll instead provide a dedicated "plain text"
input mode, along with utilities to decode from and encode into this new
text mode, that achieves a "forced" text representation of the field
value but still typed.